fix: sharing of prompts and what a user sees (#40)
* - enabled Prompts for users
- prevented users from editing and deleting

* fix: update community sharing default to false and remove share functionality from prompts

* feat: add public access tooltip and translations for prompts

* fix: clean up unused code and restore missing translation for delete prompt error

---------

+++ src/lib/components/workspace/Prompts.svelte
@@ -39,23 +39,13 @@
 	let filteredItems = [];
 	$: filteredItems = prompts.filter((p) => query === '' || p.command.includes(query));
 
-	const shareHandler = async (prompt) => {
-		toast.success($i18n.t('Redirecting you to OpenWebUI Community'));
+	let isAdmin = false;
 
-		const url = 'https://openwebui.com';
-
-		const tab = await window.open(`${url}/prompts/create`, '_blank');
-		window.addEventListener(
-			'message',
-			(event) => {
-				if (event.origin !== url) return;
-				if (event.data === 'loaded') {
-					tab.postMessage(JSON.stringify(prompt), '*');
-				}
-			},
-			false
-		);
-	};
+	onMount(async () => {
+		await init();
+		loaded = true;
+		isAdmin = $user?.role === 'admin';
+	});
 
 	const cloneHandler = async (prompt) => {
 		sessionStorage.prompt = JSON.stringify(prompt);
@@ -70,6 +60,11 @@
 	};
 
 	const deleteHandler = async (prompt) => {
+		if (!isAdmin) {
+			toast.error($i18n.t('delete_prompt_error'));
+			return;
+		}
+
 		const command = prompt.command;
 		await deletePromptByCommand(localStorage.token, command);
 		await init();
@@ -79,11 +74,6 @@
 		prompts = await getPromptList(localStorage.token);
 		await _prompts.set(await getPrompts(localStorage.token));
 	};
-
-	onMount(async () => {
-		await init();
-		loaded = true;
-	});
 </script>
 
 <svelte:head>
@@ -154,19 +144,27 @@
 						</div>
 
 						<div class=" text-xs px-0.5">
-							<Tooltip
-								content={prompt?.user?.email ?? $i18n.t('Deleted User')}
-								className="flex shrink-0"
-								placement="top-start"
-							>
-								<div class="shrink-0 text-gray-500">
-									{$i18n.t('By {{name}}', {
-										name: capitalizeFirstLetter(
-											prompt?.user?.name ?? prompt?.user?.email ?? $i18n.t('Deleted User')
-										)
-									})}
-								</div>
-							</Tooltip>
+							{#if prompt.access_control == null}
+								<Tooltip content="public" className="flex shrink-0" placement="top-start">
+									<div class="shrink-0 text-gray-500">
+										{$i18n.t('Public')}
+									</div>
+								</Tooltip>
+							{:else if prompt?.user?.role === 'admin' || (prompt?.user?.role === 'user' && prompt.access_control != null)}
+								<Tooltip
+									content={prompt?.user?.email ?? $i18n.t('Deleted User')}
+									className="flex shrink-0"
+									placement="top-start"
+								>
+									<div class="shrink-0 text-gray-500">
+										{$i18n.t('By {{name}}', {
+											name: capitalizeFirstLetter(
+												prompt?.user?.name ?? prompt?.user?.email ?? $i18n.t('Deleted User')
+											)
+										})}
+									</div>
+								</Tooltip>
+							{/if}
 						</div>
 					</a>
 				</div>
@@ -193,9 +191,6 @@
 					</a>
 
 					<PromptMenu
-						shareHandler={() => {
-							shareHandler(prompt);
-						}}
 						cloneHandler={() => {
 							cloneHandler(prompt);
 						}}
